+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Ability to execute functions asynchronously via `sll_execute_function`
 - Allocator API: `sll_allocator_from_memory`, `sll_allocator_init`, `sll_allocator_move`, `sll_allocator_release` and `sll_allocator_resize` ([#224])
 - Implemented `input.sll`, `input$char`, `input$int` and `input$string` ([#244])
+- Implemented `sll_cpu_t`, `SLL_CPU_ANY`, `sll_platform_current_thread` and `sll_platform_set_cpu` ([#243])
 - Implemented `sll_set_sandbox_flag_string` and `sys$set_sandbox_flags`
 
 ### Changed